Merge apps and add distributed

+++ build/test.sh
@@ -1,9 +1,18 @@
 NS3_VERSION="ns-3.30.1"
 
+# Usage help
+if [ "$1" == "--help" ]; then
+  echo "Usage: bash test.sh [--help, --no_coverage]"
+  exit 0
+fi
+
 # Test results output folder
 rm -rf test_results
 mkdir -p test_results
 cd ${NS3_VERSION} || exit 1
+
+# Rebuild to have the most up-to-date
+bash rebuild.sh
 
 # Empty coverage counters
 echo "Zeroing coverage counters"
@@ -11,7 +20,13 @@ lcov --directory build/debug_all --zerocounters
 
 # Perform coverage test
 echo "Performing tests for coverage"
-python test.py -v -s "basic-sim" -t ../test_results/test_results_basic_sim
+python test.py -v -s "basic-sim-core" -t ../test_results/test_results_core
+python test.py -v -s "basic-sim-apps" -t ../test_results/test_results_apps
+
+# Stop if we don't want a coverage report
+if [ "$1" == "--no_coverage" ]; then
+  exit 0
+fi
 
 # Back to build/ directory
 cd .. || exit 1
@@ -32,4 +47,5 @@ echo "Coverage report is located at: coverage_report/index.html"
 
 # Show results
 echo "Display test results"
-cat test_results/test_results_basic_sim.txt
+cat test_results/test_results_core.txt
+cat test_results/test_results_apps.txt
